Implemented a basic map loader
Created a basic map loader script and a map defaults script.
Currently the scene only loads a set location but it is in a good place, ready for further development.

The location read by the location service manager is to be read by the map loader as the map location