Animals: Roam
An Animals currently roam around an anchor point, moving to a space close to themselves. For pets the anchor point is the player GameObject, for non-pet animals it will be there spawn location.

NOTE: Pets need to roam the full home area.
NOTE: The player should be able to attach pets onto a lead.

- Finished first draft of animal roaming
- Updated animal stats amount
- Neatened animal canvas